argparse test: Tighten regex against travis' shenanigans
Travis puts the commit message in an environment variable, so if it
contains the string `_flag` this would match TRAVIS_COMMIT_MESSAGE.

That happened in ca91c201c35448b953da154d01d55f149aa2419c, so the
tests failed.

We simply tighten the regex a little more, and make a commit message
that doesn't include the string.