Allow eval to see the tty if its output is not piped
Commit 5fccfd83ece8a3126f6dc65be7b85aefc7424f14, with the fix for #6806,
switched eval to buffer its output (like other builtins do). But this
prevents using eval with commands that wants to see the tty, especially
fzf. So only buffer the output if the output is piped to the next process.

This will solve #6955 (which needs to go into a point release).