fish_config: 'Webify' color definitions
The colors defined in `colorutils.js` lack the leading
`#` character and do not fully follow the html/css spec
(https://www.w3.org/TR/css-color-4/#typedef-hex-color).

While this (somehow) works for the most part, a few
colors get lost along the way and do not display in
the customization selector nor in the preview when
selected.

To fix this, let's prepend the missing '#' character
to all colors defined in `colorutils` and adjust a
few other places that were already doing that (as to
avoid doubling up the '#'s).