Add fc(1posix) - process the command history list
This is useful to combine and rerun several commands.
For example having the last 4 commands in $history:

$ string join \n $history[1..4]
./test
git clean -dfx
ls
make

$ fc -4 -1

opens an editor where you can drop the ls, maybe move the git command to
the start, join all lines with '; and' and save. Now the prompt contains
one command, ready to be executed.

Note that the numbering is inverted to $history to comply with POSIX.

The implementation is based on share/functions/edit_command_buffer.fish.

fc is specified by POSIX in

https://pubs.opengroup.org/onlinepubs/9699919799/utilities/fc.html

For now only the first and last options are implemented.

Fixes #5030